fix: Fixed comments UI
