+++ apps/dashboard/src/pages/index.jsx
@@ -9,7 +9,6 @@ import Newsletter from '../components/index/newsletter';
 import Footer from '../components/index/footer';
 import {useEffect} from 'react';
 import sdk from '@lettercms/sdk';
-import Layout from '../components/tracingLayout';
 import {getSession} from 'next-auth/react';
 import {useToken} from '@/lib/userContext';
 import jwt from 'jsonwebtoken';
@@ -25,8 +24,7 @@ export default function Home({referrer, isAdmin}) {
     }
   }, [accessToken, isAdmin, referrer]);
 
-  return <Layout>
-    <div>
+  return <div>
       <Head/>
       <Header />
       <Description/>
@@ -36,8 +34,7 @@ export default function Home({referrer, isAdmin}) {
       <Newsletter/>
       <Video/>
       <Footer/>
-    </div>
-  </Layout>;
+    </div>;
 }
 
 export async function getStaticProps({req, res}) {
